+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fire "changed:" events after updates have been fully settled with the memory source
+- Fix test for multiple movement locations/geometries and confirm that no fetch calls occur when offline
 
 ## [0.10.0] - 2023-02-05